+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Changelog
 
+## 0.20.20
+
+* `Repos`:
+    * `Exposed`:
+        * Add opportunity for setup flows in `AbstractExposedCRUDRepo`
+
 ## 0.20.19
 
 * `Versions`: